requirements.txt
openclaw>=0.1.0
pyyaml>=5.4.1
colorama>=0.4.4
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
